demo: vite env proof of concept
This commit demonstrates one workaround for getting access to the
dotenv files exposed to vite.

It works by adapting vite's `defineConfig`, and passing a callback into
it and extracting the `mode` parameter passed to the callback. `mode`'s
value is either `development` or `production`, and we want to pick an
`.env` file depending on the value of `mode`.

One way to do this--as demonstrated in this proof of concept--is to call `loadEnv()` with our `mode` variable, and use the returned value to define a global on the `window` which will be accessible to the app in the browser.

In a browser session, logging window.__MODE__ will give you the mode the
app is executing as, and window.__ENV__ will give you the environment
variables defined in the corresponding .env file (either
.env.development or .env.production).