fix(merchant): listings view better accomodates small-medium viewports
this change fixes products being squashed horizontally for viewports
that are medium-small size. the squashing happened due to the grid being
split into three parts before the before the mobile-first
one-product-per-row breakpoint has been hit. the issue could be seen
when viewing a shop in a browser window that was only half the screen.

with this change there's a default set as one-product per row, which
then goes into a grid of two for small viewports and continues on this
two-grid until hitting the breakpoint for larger viewports, which is
when the grid is divide into parts of three instead.