various ui fixes
-------------
ui(merchant): top left logo display pointer on hover
-------------

The top left element is used as a component to lead a user to the back
to /listings when clicked. We now put cursor: pointer on it, so that a
cursor turns into a pointing hand to signify "hey i am clickable!".

--------------------
fix(listing-detail): correctly update cart quantity on add to basket
--------------------

This commit changes quantity in ListingDetail.tsx to be of type number
rather than string. We also set defaults of values to be 1 rather than
empty string or 0, as it is a better UX for customers to be able to click
"Add to basket" rather than having to increment first and then click
Add.

The major change concerns what to do when "Add to basket" is clicked for
an item that is already in the cart. Previously the item quantity was fully
replaced with the newly added quantity. A more intuitive behaviour is to
instead add the two quantities such that:

cart(item).quantity = cart(item).quantity + newQuantity

Also, a little util is introduced in the form of say(), which is like
warn but also includes the actual function name as context.

---------------------
test(listing-detail): adjust test to new behaviour
---------------------

also threw in some minor refactors to listing-detail, using parseInt and
trimming away my describe function for now in preference of the `debug`
convention used elsewhere in the codebase.

main portion of the commit is that the test has been updated to reflect
the new behaviour

--------------
fix(nav/cart): basket icon is correctly reset after successful checkout
--------------

In the hook `useCurrentOrder`, we "reset" `currentOrder` for the
following states:
* OrderState.Canceled
* OrderState.Paid.

This resetting is done by setting currentOrder to null, causing this
bug.

The bug raises the question of using a sentinel value, instead of null,
to represent reset orders where we would do e.g.

currentOrder = SENTINEL_ORDER

where SENTINEL_ORDER is a fake order set as a constant and checked
against in addition to simple null checks.

--------------
fix(nav/menu): make entire menu row clickable, not just text
--------------