fix(listing-detail): correctly update cart quantity on add to basket
This commit changes quantity in ListingDetail.tsx to be of type number
rather than string. We also set defaults of values to be 1 rather than
empty string or 0, as it is a better UX for customers to be able to click
"Add to basket" rather than having to increment first and then click
Add.

The major change concerns what to do when "Add to basket" is clicked for
an item that is already in the cart. Previously the item quantity was fully
replaced with the newly added quantity. A more intuitive behaviour is to
instead add the two quantities such that:

cart(item).quantity = cart(item).quantity + newQuantity

Also, a little util is introduced in the form of say(), which is like
warn but also includes the actual function name as context.