fix(nav/cart): basket icon is correctly reset after successful checkout; entire menu row is now clickable (#346)
* fix(nav/cart): basket icon is correctly reset after successful checkout

In the hook `useCurrentOrder`, we "reset" `currentOrder` for the
following states:
* OrderState.Canceled
* OrderState.Paid.

This resetting is done by setting currentOrder to null, causing this
bug.

The bug raises the question of using a sentinel value, instead of null,
to represent reset orders where we would do e.g.

currentOrder = SENTINEL_ORDER

where SENTINEL_ORDER is a fake order set as a constant and checked
against in addition to simple null checks.

* fix(nav/menu): make entire menu row clickable, not just text