fix(modal): click outside menu consistently closes modal
The issue was that width was being used to adjust the positioning for
the cart dropdown. This caused the cart dropdown to billow out invisibly
and absorb clicks that were thought to be outside of the menu and thus
going counter to expectations that the modal would be closed.

Ideally, the dropdown would be positioned in the component hierarchy
such that it could simply be aligned to the end and we don't need any of
these margin-style adjustments, but this was a less invasive fix :)

closes #375