remerging *write tag/commit to release-info.json* & other bugs from user test (#396)
* write tag/commit to release-info.json

* success toast in listing detail should not bump up the UI

* disable button (#391)

* fix(modal): click outside menu consistently closes modal (#376)

The issue was that width was being used to adjust the positioning for
the cart dropdown. This caused the cart dropdown to billow out invisibly
and absorb clicks that were thought to be outside of the menu and thus
going counter to expectations that the modal would be closed.

Ideally, the dropdown would be positioned in the component hierarchy
such that it could simply be aligned to the end and we don't need any of
these margin-style adjustments, but this was a less invasive fix :)

closes #375

* separate CheckoutFlow to separate routes

checkout tests pass again

---------

Co-authored-by: Henry <henry@mass.market>
Co-authored-by: Al <alp@mass.market>