chore(.env): implement solution to load .env files for dev/test/prod (#418)
* chore(.env): implement solution to load .env files for dev/test/prod

This was a tricky one.

.env (aka dotenv) can be used to store configuration variables in a way
that can live outside of a repo and vary depending on whether a project
is being run in development mode (.env.development) or in production
(.env.production).

Vite's `import.meta.env` is supposed to contain the values of the
corresponding env file (.env.development for dev). But in Deno, that
doesn't work: `import.meta.env` can't be accessed.

One work-around then is to load the .env files in Vite's config during
startup and make them available on a variable that will contain the
values thanks to Vite hooking into esbuild: https://esbuild.github.io/api/#define

However, this behaves differently when running a development build with
`deno run dev` and `deno build`, so that needs to be captured.

Additionally, the testing environment does not go through Vite, and so
the corresponding variables containing the .env values are not
available. So these have to be read from disk and made available, and
done so in a way that doesn't cause a build to fail because of fs not
being available in the browser.

Everything is compounded by the iteration speed of the current project,
with multi-minute loads with Vite currently.
---

Removed env-production.ts and env-vite.ts since these were workarounds
for the .env problem.

* fix production task

* fix(.env): use relay endpoint at v4

* fix(env.ts): expound on using __ENV__ directly during prod

Added a comment to provide context on the different access of
__ENV__ when making production builds.

---------

Co-authored-by: Henry <henry@mass.market>