fix (ui/customer-listings-overview): align text and img width
Previously, the product image and the text were using different widths.
Since each "pill"-- displaying a product which can be clicked on to view
details -- can be viewed as a conceptual component, I imposed a width on
the parent div and added a full width to each of its children (img and text
portions).

Before this change, there was a bit of inteface jank and misalignments
visible for medium sizes of the viewport before snapping to the small
viewport's breakpoint.