fix (ui/cart): display cart errors at bottom not top
Previously the cart error was displayed at the top. This is not good
because it displaces all the other buttons in the interface. It's now
placed at the bottom, and doesn't jiggle around the other interactive
elements when it appears. It has also been confirmed to still be visible
even when the cart has many items.

Ideally, toasters are located in a place relative to the nav rather than
in components and "pop up" and float away, instead. But here's a quicker
fix in the meantime.