Fix minor interface problems (#430)
* fix (ui/basket): correctly align open basket with med->sm breakpoint

Prior to this change, the basket went partially out of screen before the
view snapped to the smaller viewport breakpoint.

* fix(ui/cart): opened cart dropdown is now scrollable

Previously the cart dropdown would not have a fixed height, and would
extend outside of the bottom of the viewport if very many items were
present. Now, the height adapts to the viewport and adds a scrollbar if
the combined height of the items exceeds the viewport.

* fix (ui/customer-listings-overview): align text and img width

Previously, the product image and the text were using different widths.
Since each "pill"-- displaying a product which can be clicked on to view
details -- can be viewed as a conceptual component, I imposed a width on
the parent div and added a full width to each of its children (img and text
portions).

Before this change, there was a bit of inteface jank and misalignments
visible for medium sizes of the viewport before snapping to the small
viewport's breakpoint.

* fix (ui/customer-listings-overview): use suitable line-height

The previous line-height looked cramped on some interfaces for
multi-line product titles. This change creates a bit more space.

* fix (ui/cart): display cart errors at bottom not top

Previously the cart error was displayed at the top. This is not good
because it displaces all the other buttons in the interface. It's now
placed at the bottom, and doesn't jiggle around the other interactive
elements when it appears. It has also been confirmed to still be visible
even when the cart has many items.

Ideally, toasters are located in a place relative to the nav rather than
in components and "pop up" and float away, instead. But here's a quicker
fix in the meantime.