revert: revert changes to frontend pipeline
Reverts parts of 50403f9071c5f6715240fcf3fcdb713f274dbae3

The reason is the frontend on my machine would not successfully execute
with the reverted changes in place; the shop ui would never load.

In order to progress on #436, I needed to have a working baseline from
which to start making changes.