fix(share): fix hardcoded share link (#474)
This uses window.location to derive both the protocol (so http: and
https: is chosen correctly, depending on the url) and the actual domain
that is visited, instead of being hardcoded to demo-shop.

However the BigInt routine that converts the hex into base10 seems
broken on my end, at least locally.

closes #400 #469