fix(share): populate share link using url params
For reasons seemingly related to external libs like viem, useShopId has
to operate on a base-10 version of the shop id. This however doesn't
function well in the Share view. Thus this commit, which tries to source
the shop id from either the url params or the env file, before
defaulting to the empty string.

closes #476