refactor (err): introduce utils/getErrLogger(logger, setErrorMsg?) -> logError(msg, error?)
This commit refactors the error handling based on patterns already in the codebase.

The refactor is done by introducing the getErrorLogger function inside
of `utils`. This function accepts a logtape logger, and optionally a
react state setter function for setting an error message in the
interface. It returns a function, logError, that accepts and error message, and
optionally an error.

`logError` encapsulates how we want to treat errors:

If the logged error has an error, output it into the console with the
error object first, and then log the error message. If the state setter
function is present, then we call that with the passed-in error message.