Improve error handling (#552)
* test(logtape/frontend): use `lowestLevel` not `level`

part of #551

* test(pay): for errors - surface nicer messages for the user and error details for the logs

progresses #537

* test(error handling): improve error handling and unify logger naming

* getLogger was setting names alternating between kebab-case and
  PascalCase. We now use PascalCase - or rather we used the file name the logger is
  declared within but skip the .tsx/.ext.
* Progressed #537 by making the error toast only use strings instead of
  sometimes including error output.

* test(logtape): config logtape to be less noisy in glitch tip

closes #551

* tooling(gitignore): make sure glitch.ts is not added

* refactor (err): introduce utils/getErrLogger(logger, setErrorMsg?) -> logError(msg, error?)

This commit refactors the error handling based on patterns already in the codebase.

The refactor is done by introducing the getErrorLogger function inside
of `utils`. This function accepts a logtape logger, and optionally a
react state setter function for setting an error message in the
interface. It returns a function, logError, that accepts and error message, and
optionally an error.

`logError` encapsulates how we want to treat errors:

If the logged error has an error, output it into the console with the
error object first, and then log the error message. If the state setter
function is present, then we call that with the passed-in error message.

* Update packages/frontend/src/utils/helper.ts

Co-authored-by: Henry <111202+cryptix@users.noreply.github.com>

---------

Co-authored-by: Henry <111202+cryptix@users.noreply.github.com>