wip error handling
This commit demonstrates how to use a context object attached to a
logtape handler, such that the keys of the context object can be used to
fluff up error traces before they hit glitchtip using the Sentry API's
`setTag`.

See ListingDetail.tsx for how the logger is set up initially, and
logtape-sentry-sink/mod.ts for the usage of setTag.

We also set up a list of predefined keys, limiting the invocation of
setTag to those keys. These keys be data that is useful for debugging
crash reports, such as an orderId.