fix logtape configuration conflict
anything lower than warning was not firing in the local dev environment
since the change in 0f6fec3782153565d5c74f116b4e22bcbb52d02e

this commit addresses the issue by choosing the lowest logging level
depending on the environment. it also makes sure glitchtip doesn't
receive logging events if the environment is set to hardhat (local
development).

glitchtip was receiving events because the `VITE_SENTRY_DSN` was set in
`.env` but not in `.env.development`, which caused vite on my local
machine to pull in the `VITE_SENTRY_DSN` key since seemingly
`.env.development` acts as an override/additional source of keys in
addition to what is set in `.env`.