test(EditListing): fix test flakiness
The test in this commit was flaking due to an unrelated reason for what
was actually being tested.

Previous to this commit, the test was clearing the stock input from `123` and then
typing in `321`. Occasionally this flaked: the clear was registered but
the subsequent change event contained `1233` i.e. the previous value
plus the newly typed '3'.

Rather than spend more time digging into the exact cause, which seemed
to be purely timing oriented and could be related to the happyDom
mocking environment, I altered the test in a way that sidesteps the
issue behind the flake while still testing the relevant concern: making
sure stock amount is updated correctly in stateManager and that it
the new stock value is reflected in the interface.