logging: final touches and configuration; use setContext
The logger behaves in the following way:

If sentry is enabled (a DSN is set) and the chain is set to something
other than the dev environment (hardhat), then:

* logs with log levels lower than warning will not be logged (either to
  console or to sentry)
* all logs of level warning or error will create events in glitch tip

if sentry is not enabled:

* the lowest log level is debug
* events will not be sent to glitchtip

Additionally, any objects that are logged (e.g. with syntax
logger.warn("message", { myObject })) will be exposed in glitchtip using
the Sentry API setContext. It will be displayed as a key-value row under
the header of "debug context info". If `logger.error` was used, then
the header instead is `thrown error`.