misc (logging): enable browser-side debug logs if query param "&mass-debug=true" is set
This came about from discussion in https://github.com/masslbs/Tennessine/pull/576#pullrequestreview-2900482058

The feature in question allows setting a query parameter and reloading
the page to enable selective debug logs in the browser.

The query parameter to be set is "mass-debug=true". If set, the logger
changes its lowest level to "debug". This behaviour is currently only
enabled for the relay client, `packages/client`.

For posterity, here is the rationale for doing it in the way it is
implemented:

"the easy way to do this would have as a consequence that the sentry sink
is disabled when ?mass-debug="true" is set. turning off the sentry sink
is a tradeoff to keep the logtape config from becoming really
complicated and also to prevent spamming glitchtip.

by the way: if you're specifically/primarily after logging the relay
client's debug logs, then the solution is a lot simpler than it would be
for turning on all debug logs! in that case we could selectively set
just the relay client logs to have lowestLevel: "debug", and leave all
the other settings (e.g. the sentry sink) intact"