Use Sentry's setTag, setContext and improve general logging behaviour (#576)
* wip error handling

This commit demonstrates how to use a context object attached to a
logtape handler, such that the keys of the context object can be used to
fluff up error traces before they hit glitchtip using the Sentry API's
`setTag`.

See ListingDetail.tsx for how the logger is set up initially, and
logtape-sentry-sink/mod.ts for the usage of setTag.

We also set up a list of predefined keys, limiting the invocation of
setTag to those keys. These keys be data that is useful for debugging
crash reports, such as an orderId.

* fix logtape configuration conflict

anything lower than warning was not firing in the local dev environment
since the change in 0f6fec3782153565d5c74f116b4e22bcbb52d02e

this commit addresses the issue by choosing the lowest logging level
depending on the environment. it also makes sure glitchtip doesn't
receive logging events if the environment is set to hardhat (local
development).

glitchtip was receiving events because the `VITE_SENTRY_DSN` was set in
`.env` but not in `.env.development`, which caused vite on my local
machine to pull in the `VITE_SENTRY_DSN` key since seemingly
`.env.development` acts as an override/additional source of keys in
addition to what is set in `.env`.

* fix(.env): remove DSN

* fix(tests): enable logging in frontend tests

following the pattern established in `packages/client/mod_test.ts`, this commit re-enables testing output by
importing the default logtape logger at "@massmarket/utils/logger"

* test(logging): set env MAX_LOGS to log tests

to enable the full suite of debug logging, run the frontend tests as:

MAX_LOGS=1 deno test -A

* test(EditListing): fix test flakiness

The test in this commit was flaking due to an unrelated reason for what
was actually being tested.

Previous to this commit, the test was clearing the stock input from `123` and then
typing in `321`. Occasionally this flaked: the clear was registered but
the subsequent change event contained `1233` i.e. the previous value
plus the newly typed '3'.

Rather than spend more time digging into the exact cause, which seemed
to be purely timing oriented and could be related to the happyDom
mocking environment, I altered the test in a way that sidesteps the
issue behind the flake while still testing the relevant concern: making
sure stock amount is updated correctly in stateManager and that it
the new stock value is reflected in the interface.

* use sync configure for logger

* test(components): enrich loggers across non-hooks comps

Crash loggers now include non-PII information like orderId, keycard
public key, shopId. Which are included depends on the individual
component.

* logging: final touches and configuration; use setContext

The logger behaves in the following way:

If sentry is enabled (a DSN is set) and the chain is set to something
other than the dev environment (hardhat), then:

* logs with log levels lower than warning will not be logged (either to
  console or to sentry)
* all logs of level warning or error will create events in glitch tip

if sentry is not enabled:

* the lowest log level is debug
* events will not be sent to glitchtip

Additionally, any objects that are logged (e.g. with syntax
logger.warn("message", { myObject })) will be exposed in glitchtip using
the Sentry API setContext. It will be displayed as a key-value row under
the header of "debug context info". If `logger.error` was used, then
the header instead is `thrown error`.

* misc (logging): enable browser-side debug logs if query param "&mass-debug=true" is set

This came about from discussion in https://github.com/masslbs/Tennessine/pull/576#pullrequestreview-2900482058

The feature in question allows setting a query parameter and reloading
the page to enable selective debug logs in the browser.

The query parameter to be set is "mass-debug=true". If set, the logger
changes its lowest level to "debug". This behaviour is currently only
enabled for the relay client, `packages/client`.

For posterity, here is the rationale for doing it in the way it is
implemented:

"the easy way to do this would have as a consequence that the sentry sink
is disabled when ?mass-debug="true" is set. turning off the sentry sink
is a tradeoff to keep the logtape config from becoming really
complicated and also to prevent spamming glitchtip.

by the way: if you're specifically/primarily after logging the relay
client's debug logs, then the solution is a lot simpler than it would be
for turning on all debug logs! in that case we could selectively set
just the relay client logs to have lowestLevel: "debug", and leave all
the other settings (e.g. the sentry sink) intact"