feat: catch relay auth error and throw up modal blocking interaction
This commit catches the error that originates from `/packages/client`
on calling `authencation` when another session is already in process.

Previously this situation would not be signaled to the user, which was
problematic as any subsequent interactions (e.g. cart changes) would not
be saved due to not having a valid connection with the relay.

closes #411
related #385 #583