feat(shops-overview): root route displays a placeholder overview
This commit changes the behaviour of the root route of the frontend to
display the Shops overview described by the Shops overview epic in
Anytype and issue #639.

It sketches out some of the necessary parts, but it is by no means
complete.

Also the layout and interactive elements should cohere better with the
rest of the application, and changes might have to be made to the
navigation to accomodate the shops overview (is this view entirely
standalone? should the cart be visible in top right?)

Some outstanding questions:

* What image dimensions are the suggested ones for shops?
* How long description should be the suggested default before it is
  trimmed?
* What is the suggested max length of a shop name before it is trimmed?
* How should we sort the shops overview: by shopID ('more random'), by
  alphanumerical sort of the shop name? or by some other characteristic

Additionally:

* Actual shop data needs to be pulled in via contracts, which will be
  possible to illustrate after the shop generator of #587 is merged
  (this PR should be rebased onto main once that happens)